Fix miscellaneous size_t truncations
Bug: 516

+++ crypto/fipsmodule/ec/p256-nistz_test.cc
@@ -42,7 +42,7 @@
 TEST(P256_NistzTest, SelectW5) {
   // Fill a table with some garbage input.
   alignas(64) P256_POINT table[16];
-  for (size_t i = 0; i < 16; i++) {
+  for (uint8_t i = 0; i < 16; i++) {
     OPENSSL_memset(table[i].X, 3 * i, sizeof(table[i].X));
     OPENSSL_memset(table[i].Y, 3 * i + 1, sizeof(table[i].Y));
     OPENSSL_memset(table[i].Z, 3 * i + 2, sizeof(table[i].Z));
@@ -72,7 +72,7 @@ TEST(P256_NistzTest, SelectW5) {
 TEST(P256_NistzTest, SelectW7) {
   // Fill a table with some garbage input.
   alignas(64) P256_POINT_AFFINE table[64];
-  for (size_t i = 0; i < 64; i++) {
+  for (uint8_t i = 0; i < 64; i++) {
     OPENSSL_memset(table[i].X, 2 * i, sizeof(table[i].X));
     OPENSSL_memset(table[i].Y, 2 * i + 1, sizeof(table[i].Y));
   }
